refactor: Resolves incorrect drawable resource mapping and extracts domain logic into use cases.
